feat(Knowledge Graph Comp): graph visualizer_adressing comments
