+++ packages/react-components/src/graph-visualizer/graph.ts
@@ -14,12 +14,17 @@ import type {
 // `cytoscape-cise` does not have TypeScript definitions.
 //@ts-ignore
 import cise from 'cytoscape-cise';
-import type { SetRequired, ValueOf } from 'type-fest';
-
-import type { Health } from './lib/types';
+import type { ValueOf } from 'type-fest';
+import {
+  EDGE_DEFAULT_STYLE_PROPS,
+  EDGE_HOVER_STYLE_PROPS,
+  EDGE_SELECTED_STYLE_PROPS,
+  NODE_DEFAULT_STYLE_PROPS,
+  NODE_HOVER_STYLE_PROPS,
+  NODE_SELECTED_STYLE_PROPS,
+} from './defaultProps';
 import { isNil, isString } from './lib/utils/lang';
-import { GRAPH_COLORS } from './constants';
-import { createSvgString, getPolygonPoints, getSvgSize } from './svg';
+import { getPolygonPoints, getSvgSize } from './svg';
 import type {
   EdgeArrowStyleProps,
   EdgeData,
@@ -27,7 +32,6 @@ import type {
   EdgeStyleProps,
   EventDetail,
   NodeData,
-  NodeRenderData,
   NodeStyleProps,
   Subscriber,
 } from './types';
@@ -36,58 +40,6 @@ import type {
 interface NodesDefinition {
   nodes: NodeDefinition[];
 }
-
-const EDGE_DEFAULT_STYLE_PROPS: SetRequired<EdgeStyleProps, 'targetArrow'> = {
-  color: GRAPH_COLORS.GRAY_40,
-  targetArrow: {
-    color: GRAPH_COLORS.GRAY_40,
-    scale: 1.5,
-    shape: 'triangle',
-  },
-  width: 3,
-};
-
-const EDGE_HOVER_STYLE_PROPS: EdgeStyleProps = {
-  ...EDGE_DEFAULT_STYLE_PROPS,
-  color: GRAPH_COLORS.GRAY_10,
-  targetArrow: {
-    ...EDGE_DEFAULT_STYLE_PROPS.targetArrow,
-    color: GRAPH_COLORS.GRAY_10,
-  },
-  width: 3,
-};
-
-const EDGE_SELECTED_STYLE_PROPS: EdgeStyleProps = {
-  ...EDGE_HOVER_STYLE_PROPS,
-  width: 6,
-};
-
-const NODE_DEFAULT_STYLE_PROPS: NodeStyleProps = {
-  backgroundImage: getNormalSvg,
-  borderWidth: 3,
-  labelTextColor: GRAPH_COLORS.GRAY_14,
-  labelTextSize: 16,
-  labelOffset: 12,
-  labelPadding: 8,
-  labelPosition: 'bottom',
-  shape: getNodeShape,
-  transitionDuration: 300,
-};
-
-const NODE_HOVER_STYLE_PROPS: NodeStyleProps = {
-  ...NODE_DEFAULT_STYLE_PROPS,
-  backgroundImage: getHoverSvg,
-  borderColor: GRAPH_COLORS.WHITE,
-  labelTextColor: GRAPH_COLORS.WHITE,
-};
-
-const NODE_SELECTED_STYLE_PROPS: NodeStyleProps = {
-  ...NODE_DEFAULT_STYLE_PROPS,
-  backgroundImage: getSelectedSvg,
-  borderColor: GRAPH_COLORS.WHITE,
-  borderWidth: 6,
-  labelTextColor: GRAPH_COLORS.WHITE,
-};
 
 cytoscape.use(cise);
 
@@ -403,48 +355,6 @@ function getEdgeEndpointStyle(endpoint: EdgeEndpoint) {
   return endpoint;
 }
 
-function getHoverSvg(node: NodeSingular) {
-  let { hoverSvg, isDirty, size, shape, state } = node.data() as NodeRenderData;
-
-  if (isNil(hoverSvg) || isDirty) {
-    hoverSvg = createSvgString({
-      ...NODE_HOVER_STYLE_PROPS,
-      backgroundColor: getNodeColor(state),
-      shape,
-      size,
-    });
-    node.data({ hoverSvg: hoverSvg, isDirty: false });
-  }
-
-  return hoverSvg;
-}
-
-function getNodeColor(state: Health) {
-  switch (state) {
-    case 'critical':
-      return GRAPH_COLORS.HEALTH_CRITICAL;
-    case 'high':
-      return GRAPH_COLORS.HEALTH_HIGH;
-    case 'low':
-      return GRAPH_COLORS.HEALTH_LOW;
-    case 'medium':
-      return GRAPH_COLORS.HEALTH_MEDIUM;
-    case 'offline':
-      return GRAPH_COLORS.HEALTH_OFFLINE;
-    case 'ok':
-      return GRAPH_COLORS.HEALTH_OK;
-    default:
-      return GRAPH_COLORS.HEALTH_UNKNOWN;
-  }
-}
-
-function getNodeShape(node: NodeSingular): Css.NodeShape {
-  const { shape } = node.data() as any as NodeRenderData;
-  if (isNil(shape)) return 'ellipse';
-  if (shape === 'hexagon') return 'polygon';
-  return shape;
-}
-
 function getNodeStyle({
   backgroundImage,
   backgroundOpacity,
@@ -492,32 +402,4 @@ function getNodeStyle({
     'z-index': zIndex ?? 0,
     ...labelProps,
   };
-}
-
-function getNormalSvg(node: NodeSingular) {
-  let { isDirty, normalSvg, size, shape, state } = node.data() as NodeRenderData;
-
-  if (isNil(normalSvg) || isDirty) {
-    normalSvg = createSvgString({
-      ...NODE_DEFAULT_STYLE_PROPS,
-      backgroundColor: getNodeColor(state),
-      borderColor: getNodeColor(state),
-      shape,
-      size,
-    });
-    node.data({ normalSvg: normalSvg, isDirty: false });
-  }
-
-  return normalSvg;
-}
-
-function getSelectedSvg(node: NodeSingular) {
-  let { isDirty, selectedSvg, size, shape, state } = node.data() as NodeRenderData;
-
-  if (isNil(selectedSvg) || isDirty) {
-    selectedSvg = createSvgString({ ...NODE_SELECTED_STYLE_PROPS, backgroundColor: getNodeColor(state), shape, size });
-    node.data({ selectedSvg: selectedSvg, isDirty: false });
-  }
-
-  return selectedSvg;
 }
