+++ packages/react-components/src/graph-visualizer/lib/svgs/icons/styles.module.css
@@ -1,4 +1,4 @@
-baseStyles.svg {
-  fill: currentColor;
-  shape-rendering: geometricPrecision;
+.svg {
+  fill: currentcolor;
+  shape-rendering: geometricprecision;
 }
